Updated the folder structure, made install faster
Also made added default applicationName. Closes #4 and #5.
