+++ roc.setup.js
@@ -1,16 +1,16 @@
 const defaultPrompt = require('roc').prompt;
-const baseConfig = require('roc-web-react').baseConfig;
 
 const prompt =
     defaultPrompt.concat([{
         type: 'input',
         name: 'rocPort',
         message: 'What\'s the port for the application?',
-        default: baseConfig.port
+        default: 3000
     }, {
         type: 'input',
         name: 'rocAppName',
-        message: 'What do you what to use for <title>?'
+        message: 'What do you what to use for <title>?',
+        default: 'My Roc Application'
     }]);
 
 module.exports = {
