+++ template/app/components/clicker/style.scss
@@ -1,0 +1,4 @@
+.main {
+    background: #ccc;
+    padding: 20px;
+}
