+++ template/app/redux/clicker.js
