+++ template/app/redux/reducers.js
