+++ template/app/screens/app/style.scss
@@ -1,0 +1,4 @@
+.main {
+    background: #fff;
+    margin: auto;
+}
