+++ template/app/screens/start/style.scss
@@ -1,5 +1,5 @@
 .main {
-    background: #fff;
+    background: #ddd;
     width: 700px;
     margin: auto;
     padding: 20px;
