+++ template/roc.config.js
@@ -2,9 +2,11 @@ module.exports = {
     config: {
         applicationName: '{{{ rocAppName }}}',
         port: {{{ rocPort }}},
+        serve: ['public', 'build/client'],
+        favicon: 'favicon.png',
         build: {
-            reducers: 'reducers.js',
-            routes: 'routes.js'
+            reducers: 'app/redux/reducers.js',
+            routes: 'app/routes/index.js'
         }
     }
 };
