--- template/routes.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import { Route, IndexRoute } from 'react-router';
-
-import Main from './main';
-
-export default () => (
-    <IndexRoute component={ Main } />
-);
